+++ contrib/prune/generic_list.c
@@ -31,7 +31,7 @@ generic_list_t *new_generic_list(uint64_t size)
 	return NULL;
     }
     if (size != 0) {
-	list->data = (gl_data *) malloc(size * sizeof(gl_data));
+	list->data = (gl_data *) malloc((size_t)(size * sizeof(gl_data)));
 	if (list->data == NULL) {
 	    perror("[new_generic_list()] Error allocating memory:");
 	    return NULL;
@@ -63,7 +63,7 @@ generic_list_t *add_to_generic_list(generic_list_t * list, gl_data element)
 	    new_size = list->size * 2;
 	}
 	new_data =
-	    (gl_data *) realloc(list->data, new_size * sizeof(gl_data));
+	    (gl_data *) realloc(list->data,(size_t) (new_size * sizeof(gl_data)));
 	if (new_data == NULL) {
 	    perror("[add_to_generic_list()] Error allocating memory:");
 	    return NULL;
